remove some more old refs
